requirements.txt
fastapi
uvicorn[standard]
sqlmodel
apscheduler
jinja2
python-dotenv
python-multipart
